+++ src/components/subjects/list/index.js
@@ -1,7 +1,0 @@
-import SubjectsListComponent from './main';
-import SubjectsList from './list';
-
-export {
-  SubjectsListComponent,
-  SubjectsList,
-};