+++ src/redux/models/Category.js
@@ -1,6 +1,10 @@
 import { attr, Model } from 'redux-orm';
 
 class Category extends Model {
+  static noCategoryId = -1;
+
+  static noCategoryColor = 'gray';
+
   getShortName() {
     return this.alias || this.name;
   }
@@ -13,6 +17,7 @@ Category.fields = {
   name: attr(),
   alias: attr(),
   description: attr(),
+  color: attr(),
 };
 
 export default Category;