+++ src/shared/UI/pickers/index.js
@@ -1,9 +1,11 @@
 import CalendarPicker from './calendar';
+import ColorPicker from './color';
 import Picker from './picker';
 import ModalPicker from './modalPicker';
 
 export {
   CalendarPicker,
+  ColorPicker,
   Picker,
   ModalPicker,
 };